requirements.txt
click==7.1.2
Pillow==8.0.1
